users/tempentries: Do not expose the temporary registered groups
While it was important to expose the temporary users to NSS in the past,
this is not required anymore because they are now existing only while we
are updating an User, and during that phase the NSS database is locked,
so there's no way that there may be a race on defining the UIDs, at
least externally.

Authd has still to make sure that while registering new users (pre-auth
or not) or new groups the UIDs are unique, but this has been ensured
already in the previous commits
