+++ internal/users/tempentries/tempentries.go
@@ -186,11 +186,11 @@ func (r *TemporaryRecords) RegisterPreAuthUser(loginName string) (uid uint32, er
 	}
 }
 
-// RegisterGroupForUser registers a temporary group with a unique GID in our
-// NSS handler (in memory, not in the database) for the provided UID.
+// RegisterGroupForUser registers a temporary group with a unique GID in
+// memory for the provided UID.
 //
-// Returns the generated GID and a cleanup function that should be called to remove
-// the temporary group once the group was added to the database.
+// Returns the generated GID and a cleanup function that should be called to
+// remove the temporary group once the group was added to the database.
 func (r *TemporaryRecords) RegisterGroupForUser(uid uint32, name string) (gid uint32, cleanup func(), err error) {
 	for {
 		gid, cleanup, err := r.temporaryGroupRecords.registerGroup(name)
