--- internal/users/testdata/golden/TestGroupByIDAndName/Successfully_get_temporary_group_by_ID
@@ -1,4 +1,0 @@
-name: tempgroup1
-gid: 0
-users: []
-passwd: ""
