--- internal/users/testdata/golden/TestGroupByIDAndName/Successfully_get_temporary_group_by_name
@@ -1,4 +1,0 @@
-name: tempgroup1
-gid: 0
-users: []
-passwd: ""
