--- internal/users/testdata/golden/TestUserByIDAndName/Successfully_get_temporary_user_by_name
@@ -1,6 +1,0 @@
-name: tempuser1
-uid: 0
-gid: 0
-gecos: ""
-dir: /nonexistent
-shell: /usr/sbin/nologin
